Extract findAccount and urlWithoutCredentials for sharing with credential helper
