+++ app/src/lib/trampoline/url-without-credentials.ts
@@ -1,0 +1,6 @@
+export function urlWithoutCredentials(remoteUrl: string): string {
+  const url = new URL(remoteUrl)
+  url.username = ''
+  url.password = ''
+  return url.toString()
+}
